ai_generated_files/requirements.txt
```
requests
pillow
pytesseract
```
